refactor: extract shared session util, scope all commands to cwd
Extract duplicated `findActiveSessions()` from 5 command files into
a shared `utils/session.ts` with `findCurrentSession()` that uses
`getProjectName()` to prefer the session matching the current
working directory.

Commands fixed:
- `d3k errors` — was showing wrong project's errors
- `d3k fix` — was analyzing wrong project's logs
- `d3k crawl` — was crawling wrong project's app URL
- `d3k find-component` — was using wrong project's CDP

Also removes hardcoded `--cdp 9222` from crawl and find-component
(now handled by the auto-inject in cli.ts from the previous PR).

Net: -312 lines added, +107 lines removed (205 lines deleted).
