+++ src/commands/find-component.ts
@@ -6,56 +6,12 @@
  */
 
 import { spawnSync } from "node:child_process"
-import { existsSync, readdirSync, readFileSync } from "node:fs"
-import { homedir } from "node:os"
-import { join } from "node:path"
 import chalk from "chalk"
-
-interface Session {
-  projectName: string
-  cdpUrl?: string
-}
-
-function findActiveSessions(): Session[] {
-  const sessionDir = join(homedir(), ".d3k")
-  if (!existsSync(sessionDir)) {
-    return []
-  }
-
-  try {
-    const entries = readdirSync(sessionDir, { withFileTypes: true })
-    const sessions: Session[] = []
-
-    for (const entry of entries) {
-      if (entry.isDirectory()) {
-        const sessionFile = join(sessionDir, entry.name, "session.json")
-        if (existsSync(sessionFile)) {
-          try {
-            const content = JSON.parse(readFileSync(sessionFile, "utf-8"))
-            if (content.pid) {
-              try {
-                process.kill(content.pid, 0)
-                sessions.push(content)
-              } catch {
-                // Process not running
-              }
-            }
-          } catch {
-            // Skip invalid files
-          }
-        }
-      }
-    }
-
-    return sessions
-  } catch {
-    return []
-  }
-}
+import { findCurrentSession } from "../utils/session.js"
 
 function runAgentBrowser(args: string[]): { success: boolean; output: string } {
   try {
-    const result = spawnSync("d3k", ["agent-browser", "--cdp", "9222", ...args], {
+    const result = spawnSync("d3k", ["agent-browser", ...args], {
       encoding: "utf-8",
       timeout: 30000
     })
@@ -71,9 +27,9 @@ function runAgentBrowser(args: string[]): { success: boolean; output: string } {
 }
 
 export async function findComponent(selector: string): Promise<void> {
-  const sessions = findActiveSessions()
+  const session = findCurrentSession()
 
-  if (sessions.length === 0) {
+  if (!session) {
     console.log(chalk.red("❌ No active d3k sessions found."))
     console.log(chalk.gray("Make sure d3k is running first."))
     process.exit(1)
